+++ react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [23.17.13](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@23.17.12...@momentum-ui/react@23.17.13) (2020-10-10)
+
+**Note:** Version bump only for package @momentum-ui/react
+
+
+
+
+
 ## [23.17.12](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@23.17.11...@momentum-ui/react@23.17.12) (2020-10-03)
 
 **Note:** Version bump only for package @momentum-ui/react